pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-umbra"
version = "0.1.0"
description = "Local LLM-powered Zettelkasten pipeline for Obsidian"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Jamey O'Neill" }]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24,<3.0",
    "llama-cpp-python>=0.3.0",
    "model2vec>=0.3.0",
    "sentence-transformers>=3.0",
    "scipy>=1.10",
]

[project.scripts]
umbra = "umbra.cli:main"

[project.urls]
Homepage = "https://github.com/jimnoneill/obsidian-umbra"

[tool.setuptools.packages.find]
where = ["src"]
